pre-alpha test is 50% complete with only minor bugs I've been able to catch and fix on the way. Ive been attacked by the crazy damn ambush wolves TWICE now, but we tamed floofy at the summit and are on the way back to the cabin. I have to sleep but Im going to dream bout programming flooy

+++ Settings/TypeSpecificSettings.cs
@@ -8,7 +8,7 @@ namespace ExpandedAiFramework
     {
         //Include a field name named "Enable"!
         public abstract bool CanSpawn(BaseAi baseAi);
-        public abstract float GetSpawnWeight();
+        public abstract int GetSpawnWeight();
         protected override void OnChange(FieldInfo field, object oldValue, object newValue)
         {
             ShowSettingsIfEnabled();
